fix of sign comparison problems
Note my solution was fast, but there were other ways. a very simple solution was to use size_t instead, or unsigned long long,
but I don't know the order of magnitude of the sizes that are compared. A more complete solution is better, but it takes more time
 Changes to be committed:
	modified:   protocolStack/mac/packet-scheduler/enhanced-uplink-packet-scheduler.cpp
	modified:   protocolStack/mac/packet-scheduler/nb-fifo-uplink-packet-scheduler.cpp
	modified:   protocolStack/mac/packet-scheduler/nb-roundrobin-uplink-packet-scheduler.cpp
	modified:   protocolStack/mac/packet-scheduler/nb-roundrobin-uplink-packet-scheduler.h
	modified:   protocolStack/mac/packet-scheduler/nb-uplink-packet-scheduler.cpp
	modified:   protocolStack/mac/random-access/gnb-enhanced-random-access.cpp
	modified:   protocolStack/mac/random-access/gnb-enhanced-random-access.h

+++ src/protocolStack/mac/packet-scheduler/nb-roundrobin-uplink-packet-scheduler.cpp
@@ -48,7 +48,7 @@ nbRoundRobinUplinkPacketScheduler::nbRoundRobinUplinkPacketScheduler(GnbMacEntit
   u.m_listOfAllocatedRUs.clear();
   u.m_selectedMCS = -1;
 
-  for (int i=0; i< mac->GetDevice ()->GetPhy ()->GetBandwidthManager ()->GetUlSubChannels ().size (); i++)
+  for (unsigned long int i=0; i< mac->GetDevice ()->GetPhy ()->GetBandwidthManager ()->GetUlSubChannels ().size (); i++)
     {
       m_RUmap[0].push_back(-1);
       m_RUmap[1].push_back(-1);
@@ -222,7 +222,7 @@ DEBUG_LOG_START_1(SIM_ENV_SCHEDULER_DEBUG_RR_MAP)
       printMap();
 DEBUG_LOG_END
 
-      for (int i=0; i<m_RUmap[1].size(); i++)
+      for (unsigned long int i=0; i<m_RUmap[1].size(); i++)
         {
           if (m_RUmap[1][i] > 0)
             {
